Add invariant duplicated table row oracle
