+++ testar/src/org/testar/oracles/OracleSelection.java
@@ -42,6 +42,7 @@ public class OracleSelection {
 	private static String[] oraclePackages = {
 			"org.testar.oracles.generic.visual",
 			"org.testar.oracles.web.accessibility",
+			"org.testar.oracles.web.invariants",
 	};
 
 	public static List<String> getAvailableOracles() {
